PB-18 updated memoryDataService, renamed model-contact
